Reorganize directory and flesh out some details
It made sense to place the scripts in a separate folder in anticipation
of possible future scripts. Eventually, all the scripts could be in a
folder, and the GUI could be populated with a list box of all possible
script packages.

Note for future: make a dedicated XML parser so that any combination of
packages could be added/removed from a single master VAP file.